requirements.txt
streamlit==1.34.0
opencv-python-headless>=4.5.5.64
Pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
pyttsx3==2.90
streamlit-image-coordinates==0.1.5